src/requirement.txt
# Data processing
numpy
scipy
pandas
matplotlib
seaborn
jupyter

# Provenance
prov[dot]
powerlaw

# Spark
pyspark
graphframe

# ML packages
mllib

# Utilities
environs
click
requests



